+++ CHANGELOG.md
@@ -9,20 +9,23 @@
 - Changed recipes for IC2C Ocean Generator and Wave Generator because they are very OP.
 - Changed LESU casing blocks a lot cheaper to craft.
 - Added crushedCentrifued Ore from IC2 Extras to Type Filter.
-- Changed Magic Energy Converter slightly easier to craft.
 - Added comparator output for Quantum Tanks.
 - Removed extremely rare materials like Osmium from loot table, you can still get Iridium.
 - Removed Uranium and Plutonium from loot table cause radiation.
 - Added GTC loot to Twilight Forest loot tables.
-- Changed increased the durability of Crumble Horn, Peacock Fan, Ore Magnet, and Giant Pick from TForest like GT6.
+- Changed increased the durability of Crumble Horn, Peacock Fan, Ore Magnet, and Giant Pick from TForest like GT6 (configurable).
 - Added some processing recipes for TForest in GTC and IC2C machines.
 - Added ability for GT Scanners to be nuclear thermometers as well as EU scanners
 - Added Steam Reactor internal tanks info to GT Scanner.
 - Changed Beryllium from a fluid to a solid as it should be, recipe to convert from fluid to solid in the Centrifuge.
 - Changed the Dragon Egg Energy Siphon to a Trophy Energy Siphon which can make power from more sources (configurable).
+- Changed Magic Energy blocks now require Plutonium-239 a byproduct of using Uranium rods in a reactor.
 - Changed the explosion strength and heat output of Thorium based Reactors by factor of 0.25!
 - Added Isotopic, Near Depleted, Depleted, and Re-Enriched variants of Thorium and Plutonium Rods.
+- Added recipe for obtaining Technetium pre-fusion by centrifuging Plutonium-239.
 - Removed bloated processing recipes for blocks to dusts etc.. in IC2C machines.
+- Removed centrifuging wet coal to get some Thorium recipe.
+- Removed Ic2 Extras compat recipes so it can handle it itself.
 
 # 1.1.5
 - Added ability to add materials to GTC via Craft Tweaker thanks to Trinsdar.